refactor: code convention 반영
