+++ src/main/java/techcourse/myblog/domain/ArticleRepository.java
@@ -1,7 +1,11 @@
-package techcourse.myblog.repository;
+package techcourse.myblog.domain;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
+
+import java.awt.print.Pageable;
+
+import javafx.scene.control.Pagination;
 import techcourse.myblog.domain.Article;
 
 @Repository
